[root_presenter] virtual keyboard: plumb views and view koids
Lay some plumbing to enable focus sensitivity in the VKCoordinator.
Specifically:
1. Update VKCoordinator::RequestTypeAndVisibility() to require a
   view koid.
2. Pass a view koid into the VKController ctor.
3. Update VirtualKeyboardFidlTest::CreateControllerClient() to return a
   ViewRef for the View bound to the Controller. In a future CL, tests
   will use this ViewRef to inform the VKCoordinator that the
   Controller's View is focused.

Note that item 2 is a change from previous plans, where we were passing
the ViewRef to the VKController ctor. Resolving the ViewRef to a koid in
the VKCoordinator improves error handling: if the lookup fails, we
don't keep a useless VKController around.

While there: add some debug logging in VKController, and improve some
existing logs (adding the name of the function).

Bug: 76659
Test: fx test root_presenter_unittests
Change-Id: I3ac1bc8e2a7a944bbb62e554ee22836a41ab666e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590381
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>