[wlan][brcmfmac] Override incorrect reason code
On certain occations, we notice the firmware returning a failure
status for an assoc request, but with a success reason code. This
results in the WLAN stack incorrectly thinking that the connection
was successful.

The below mentioned bug is currently tracking this issue. In the
meantime, this will serve as a work around by overriding the reason
code with a generic error under these circumstances.

Bug: 85446
Test: fx test assoc_test_component
Change-Id: I10b6ca044427144de945550447de47d067f210b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588833
Fuchsia-Auto-Submit: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>