[kernel][vm] Correct PagerCounts
Previously it was assumed there would always be the full number of
buckets in use and so the oldest queue would be kNumPagerBacked from
the newest queue. Now that the LRU queue might get processed in
preparation for aging to have a queue ready for it this can result
in the oldest count being falsely reported as zero. This fixes by
always considering the LRU index to be the oldest.

Change-Id: I2f5fafe5fad13e27131e9b306f9600526562e004
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591061
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>