[fdio] Provide fdio_fd_transfer_or_clone()
FDIO previously provided APIs to clone an FD to obtain a new
kernel handle referring to the same underlying resource, or to
"unwrap" an FD to obtain the existing handle, but only if the FD
was not sharing it with other FDs in the process.

In some use-cases it is desirable to be able to transfer an FD
that is unshared, and to otherwise clone it. This was not possible
for callers to implement, given that fdio_fd_transfer() consumes
the FD on failure.

Provide an fdio_fd_transfer_or_clone() API to express the intent
to obtain a handle from an FD efficiently, whether by unwrapping
or cloning it.

Bug: 85712
Change-Id: I5d4a3a675958a24dee214316b2b8aa0a0699caee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589644
Fuchsia-Auto-Submit: Wez . <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
API-Review: Adam Barth <abarth@google.com>