[clkctl][aml-clk] Implement enable/disable via clkctl
For diagnostic and debugging purposes it's sometimes useful to be able
to enable and disable clocks from the command line.

This patch adds an enable/disable interface to the clock-impl interface
and adds functionality to clkctl to manipulate that interface.

Bug: 85139
Test: Started up a peripheral and disabled its clock. Observed that
the peripheral stopped sending data. Enabled the clock again and
observed that the peripheral started working again.
Change-Id: I34cc4aba1c3a45d3ddc3470f5eabe1457f1a42eb

Change-Id: I92acf22294cfa242b44dbd140e72384e333a899f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586709
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
API-Review: Christopher Anderson <cja@google.com>