[docs] Update old FIDL syntax in docs
This CL updates a couple pages that were using the old FIDL syntax:

* //docs/concepts/settings/policy/policy_concepts.md

  Updated FIDL from sdk/fidl/fuchsia.settings.policy/volume_policy.fidl.

* //docs/development/settings/using_settings.md

  Updated FIDL from //sdk/fidl/fuchsia.settings/accessibility.fidl. Also
  updated Rust code to use the renamed `connect_to_protocol`.

Ideally these should be part of the build and use {%includecode...}, but
fixing that would be a larger undertaking. For example we likely don't
want to put includecode markers directly in these SDK files.

Change-Id: I8366c601adbc10b078f2ccb843c8f6d3f3baccd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590902
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>