[kernel][test] Fix kstack-mp-sync-exec test
This change fixes a problem with kstack-mp-sync-exec test.  Prior to
this change, the two threads of the test could end up running on the
same CPU, which could cause the test to take tens of seconds to
complete because we'd have to wait and get lucky so that the
non-pinned thread would be moved to some other CPU.

Update the test to use two separate CPUs with each thread pinned to a
different one.

Change-Id: I91c0ba6456bdd6d1c526d74fe3eb3770680a2294
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590173
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>