[virtualization][vsock] Remove status param from WaitOnSend
The functions WaitOnSend and WaitOnReceive set up an async::Wait to wait
for descriptors to become available on the send/receive virtio queues.
The two functions currently accept a status parameter. The behaviour of
the function depends on this status, roughly as follows:

* If the status is ZX_OK or ZX_ERR_SHOULD_WAIT, a wait is set up. If the
  wait fails, an error status is returned, usually with a log message.

* If the input status is not one of the above or ZX_ERR_ALREADY_EXISTS,
  the wait is cancelled. A log message is printed, unless the input is
  ZX_ERR_STOP.

Update the functions to unconditionally set up the wait, and move error
handling logic to the call site.

Multiply: virtualization-core-tests, vmm_unittests
Bug: fxbug.dev/85702

Change-Id: Iaa418e97d1e8fea50ae901beb72092a493f04e56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584803
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>