[virtualization][vsock] Directly send reset packets to guests
Previously, when VirtioVsock would receive an invalid packet from the
guest it would create a new NullConnection to track the need to send
a reset back. This would then be processed next time the transmit path
(Mux) was called.

In this CL, we implement a send queue, and use this to directly send the
reset if descriptors are available. If not, instead of tracking the
reset in the form of a fake connection, we use a packet queue.

The send queue will be particularly useful in upcoming CLs where we want
to send control packets back to the guest from within a connection
without having to modify its state.

Multiply: virtualization-core-tests, vmm_unittests
Bug: fxbug.dev/85702

Change-Id: I6aa61db76db756da4c0cb55f06dc4b39210c3f2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584725
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>