[rust][rbe] wrapper: Ignore -Lnative input dirs
-Lnative directories were being expanded to their contents
by the wrapper, however:

1) the listing of files is not deterministic in the middle of a build
2) the individual files are thus not guaranteed to be content-stable

This caused rewrapper/reproxy to fail fatally, when contents of
files were unexpectedly written/changed.

It turns out, nothing in the build so far actually needs -Lnative=dir,
so we ignore this case completely, when it comes to staging uploads.

-Lnative=file still works the same as before, because those are direct
paths.

This should solve much of the remote build flakiness issues encountered.

This also has the benefit of reducing the number of input files to hash
and upload (still small compared to clang-dir though).

Bug: 86214

Change-Id: Ic38146a626ecf87044289141f0497f72a37e5213
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591288
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>