[scenic] Keep going w/o software color correction
Integration tests that run on FEMU exercise this path, and the session
teardown causes them to fail even though everything else is working
fine.

The alternative is to skip this codepath by not offering the
fuchsia.accessibility.ColorCorrection service to root_presenter.  This
is subpar because it lowers test coverage and makes the test less
realistic.

Continue to print the warning though, just avoid killing the session.

Test: flutter-embedder-test
Bug: 73984
Bug: 85765
Change-Id: I19857c35512ba4b3b0d1bbff100191cc12a3f2e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591051
Commit-Queue: David Worsham <dworsham@google.com>
Reviewed-by: Chris Fontas <cfontas@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>