[netstack3] Finer-grained IP, ICMP socket errors
Split the storage and computation of the cached information in an IP
socket from the storage and computation of the information which
comprises the definition of the socket. In the future, this will allow
us to update cached information when the state that it's based on
changes without duplicating code.

Right now, this lets us introduce new, finer-grained error types for
creating and sending on IP and ICMP sockets. In particular, it allows us
to create four buckets for errors:
- Errors which occur when computing the definition of an IP socket
  (e.g., picking a local IP address if one was not specified)
- The error which occurs when a socket with the given addresses already
  exists (currently, this is only for ICMP sockets, but will eventually
  be used for UDP and TCP sockets as well)
- Errors which occur when a socket becomes unroutable, which can happen
  either at creation time or when state is updated later which causes an
  existing socket to become unroutable
- Errors related to sending a packet (currently, just an MTU error)

Finer-grained error types increase the likelihood that consuming code
will be able to make use of error values exhaustively, handling every
possible error case.

Change-Id: I2968d5409de83fce3b17f06aa014c006485e47aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583207
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>