[debugdata] Change version to use relative counter
The following patch made counter pointers relative, but it did not bump the profile version:
https://github.com/llvm/llvm-project/commit/a1532ed27582038e2d9588108ba0fe8237f01844
Rust uses LLVM13 release branch, and the patch above is not part of that branch. A recent change is made to bump the profile version to 8 in LLVM14:
https://github.com/llvm/llvm-project/commit/24c615fa6b6b7910c8743f9044226499adfac4e6

This patch uses relative counters only for the profiles that are generated with version 8 or later, so that they won't be used for Rust profiles.

Change-Id: I425b268efc9e9cf62f1207d9378b5b412c1d5c49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590184
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Gulfem Savrun Yeniceri <gulfem@google.com>