[fuchsia.sys2] Change facets to Dictionary.
The type for the `facets` in the `ComponentDecl`
table was changed to `fuchsia.data.Dictionary`.
This was done in order to bring this type to parity
with its SDK counterpart, which uses the `Dictionary`
instead of a custom type.

Change-Id: I7630d750e142109b4959da397e856a941e03bf6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591365
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>