[netstack] Closing device closes DeviceControl
The logic of closing the DeviceControl channel when the device is closed
was not implemented correctly. The test
`device_control_closes_on_device_close` which was checking for it was
written incorrectly, and was observing a Netstack crash through the
channel closure instead of the expected DeviceControl channel closure.

Fix the logic and tighten the test.

Multiply: netstack-interfaces-admin-fidl-test: 30
Bug: 85877
Change-Id: I9465293088e5cfa00f21619ee5c4c3f2cf12e4f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591842
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>