[ffx] Create Writer type
This creates a library at ffx/lib/writer which holds a Writer struct
that will be injected into plugins to handle output. The initial use
case is to support machine output but the end goal is that all writing
done by a plugin will run through this type.

This is the first in a series of CLs that fleshes out this concept and
introduces it into the codebase.

Change-Id: If935ac5361cf86a381607d98e07b5a8bd46680e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576633
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>