[rust][clippy] Add flags to clippy script for tricium
The recipe that runs for tricium will build the clippy targets
for the set of changed files on its own. It will use this script
to perform that file to target mapping, and then when the analyzer
runs this script to fetch lints it will skip the ninja step because
the targets are already built.

Bug: 86101
Change-Id: Ifb47a53e179471aa0123f668ce90b77ecb81ab73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590634
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Joseph Ryan <josephry@google.com>