[lockdep] Fix edge cases for lockdep when used in the kernel.
When used in a kernel environment, lockdep has some edge cases which
need to be handled in a slightly special fashion.  Kernel code is
slightly special in that it can not only be interrupted at any point
in time due to an exception/interrupt, but the exception/interrupt
code can share locks with the kernel code which was executing
beforehand, which can lead to issues.

See the following for details:
http://fxb/84890
http://fxb/85289

To address these issues, the following changes have been made:

First, a PreValidate step has been added to the LockPolicy template
interface. This allows users to change the behavior for one of their
lock types to take actions immediately before validation happens as a
lock is acquired. This gives IrqSave spinlocks the ability to disable
interrupts before validation takes place, avoiding a potential race
which can lead to the detection of a false positive lockdep violation.

Second, a lock type's flags are now passed to the user supplied
SystemGetThreadLockState function.  This allows the kernel to choose a
different thread lock state based on the type of lock, avoiding the
context mix-up which lead to http://fxb/85289.

Test: new unit tests
Fixed: 84890, 85289
Change-Id: Ia6d68d196421198aeb1a0a2f73b6888be51f6142
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591514
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>