[ffx] Create machine flag and hook up with Writer
We add a `--machine <format>` flag to the top level ffx command with
json as the currently only supported format. This also adds the Writer
type into the set of things that can be injected into a plugin and hooks
up this flag to the construction of the Writer object that is owned by
the Injector.

Change-Id: Iebef73c05d1c0f920602a005125c6bbd9d67b1db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577347
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>