[ffx] Use machine flag in target list
This is a test case of the machine flag where we use it in the target
list command. I chose this because target list already has a format flag
with a json option so not much was needed within the plugin besides
grabbing the newly provided object and calling the relevant methods.
This still looks a bit rough because the Writer object is not managing
all output at this point. This gets cleaned up in a later CL.

Change-Id: I24aa68ae78c03fea9b1f22808e29214530bac95b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576636
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>