[ffx] Add line, info, and error methods to Writer
This adds methods to the Writer object which can be used to bridge the
gap between the current state of the world and the desired future where
most output is handled by producing one of a set of widgets that we
provide. The main method is `line` which is basically just println but
is tied together with the machine flag as well as a test mechanism.
Replacing calls to println with writer.line will allow for easy testing
and easy turning off of output when in machine mode.

The info and error methods write to stderr. I think we only need one of
these but I am still torn about what it call it. I think it should be
info as that is what people use it for. Real errors should be returned
as an error from the plugin so I don't think errors should be reported
inline. But information can be reported to this other stream.

This also does the clean up for target list to stop using stdout/stderr
directly and instead running everything through the writer object. This
is indicative of what it would look like in other plugins. Functionally
nothing should change from the current state.

Change-Id: I21075f6caa8c6606ed5500fa07b92b9fad520f50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576637
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>