[ffx] Add support for ffx attribute on Writer
This adds support for an attribute that the ffx_plugin proc macro
understands. The format of the attribute is: #[ffx(machine = SomeType)]
This must be on the Writer type in the plugin entry point, i.e.

    #[ffx_plugin()]
    pub async fn my_thing(
        rc_proxy: RemoteControlProxy,
        #[ffx(machine = SomeType)] writer: Writer,
        cmd: MyCommand,
    ) -> Result<()>

This implementation pulls the "SomeType" out and uses that to generate
some code for a schema for this plugin. Until the schema generation
stuff lands which is in a separate branch, this code is just printing
out this type.

Note that the code here generates functions but does not actually call
them so this should have no user visible impact and has zero impact for
plugins which do not use this attribute.

UPDATE: This generates a function called ffx_plugin_is_machine_supported
and does call it if the machine option is set. This causes plugins that
do not explicitly support the machine flag to not be called with it.

Parsing custom inert attributes like this in proc macros is a bit of a
pain so that code is a bit annoying but that is what we have to work
with. Note that this code also removes the attribute from the AST as
part of the parsing because otherwise the compiler complains about an
unknown attribute. This is standard procedure for this type of thing and
is possible because of the guarantees on the order that proc macros run.

Change-Id: Id3396ec62c1e4a9cdb13d8efc234b3f4ad28e50c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580651
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>