[brcmfmac] Fix SDIO control frame threading issue
When transmitting SDIO control frames, usually things like setting
iovars, there is a possibility that the thread waiting for the
transmission to complete will detect an error even if the transmission
went well. This happens because the atomic that indicates if the
transmission is complete or not is not set until AFTER the waiting
thread is woken up. Most of the time the thread scheduling allows the
transmitting thread to keep running until it sets the atomic but
sometimes we get unlucky and the waiting thread immediately gets to run.

Fix this by setting the variable BEFORE the waiting thread is woken up.

Bug: 86222
Test: Manual testing
Change-Id: I2a07d0d2b7bdac11d1f4bc76a6a5a5c33a7fe9ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591507
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>