[brcmfmac] Make Device::Init return a status code
Currently the brcmfmac Device class has a pure virtual method that
subclasses implement to perform initialization. This method has an
InitTxn object as the parameter and the initialization is supposed to
call Reply on this object with the resulting status code. This prevents
the part of Device that calls Init from inspecting the result of the
subclasses Init. In order to make decisions about further initialization
the Device class needs to know what the result of Init was.

This changes the Init method to return a zx_status_t instead and the
Device class will instead use that status to call Reply on the InitTxn
object. No implementation of init did anything else with the InitTxn
object at this point.

Test: fx test brcmfmac_sim_test
Change-Id: I282e39ccea74af86b539aec1e87622a2f3f5cd21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591289
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>