[fx sync-to] Don't fail if no repos are dirty
We currently assume that if `jiri runp` emits any output that means some
repos are dirty, but this is not correct, as `jiri runp` emits a warning
to stdout if the user has not yet chosen to enable or disable
analystics. So we should only fail this check if the output actually
references any dirty repos.

Bug: 68492
Change-Id: Ia3adaf7c34dc5c9248931c8ca0c89c5de49c9a8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591987
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>