[fidlgen][go] go:embed templates, also cleanup
This moves fidlgen_go to using go:embed rather than strings for
embedding templates. Because the go formatter is part of the standard
library rather than an external utility that's being called the common
interface for formatters needed to be simplified. That resulted in a few
follow-on changes that ended up simplifying the writing of files in
general.

Generally the code moves away from constructing a series of pipes to
just reading and writing whole generated files. In practice this
basically what was happening, but it was obscured by complexity in the
interfaces.

Change-Id: Ia3928983f39a139591f773fba59a32e03bdd6b15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589731
Commit-Queue: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>