[test_runners][elf] Refactor elf_test_runner.
This CL adds a ComponentLauncher trait to the elf_test_runner and makes
most of the runner's functionality available as rustc_library target.

The purpose here is to facilitate making variants of the elf_test_runner
that differ only in how the launch the component process. As an example,
the fuzz_test_runner described by RFC-0117 needs to install a startup
handle to a channel in the launched process, but is otherwise identical
to elf_test_runner. With this refactor, the fuzz_test_runner can simply
provide its own impl of ComponentLauncher.

Bug: 24740
Change-Id: I87313a50f5fc9e34a1f6489561823d5fe10eb81d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577305
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>