[kernel][uart][pl011] Hold spinlock when masking/unmasking RX/TX
For the pl011 masking/unmasking RX/TX interrupts is not atomic because
it requires a register read, an OR or AND, and register write.
Because it's not naturally atomic it's possible for one thread to
stomp another's changes.

Change the pl011 uart driver to guard mask/unmask operations using the
uart spinlock.

Bug: 80933
Change-Id: I44ce4dd1d6fd7865588fdbcb6d6cd6945a243b9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591517
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>