[display][i915] Use correct link rate for BIOS-initialized eDP
When a DisplayPort display gets initialized it always selects the
highest supported per-lane link rate and uses that for pixel clock
calculations. This doesn't accurate reflect the hardware refresh rate
when the panel is initialized by the bootloader (such as in the case of
eDP) and DDI & DPLL initialization is explicitly skipped in favor of the
bootloader parameters.

The atlas eDP panel runs with link clock-rate of 2160Mbps on 1080p and
4320Mbps on 4K, both of which happen to be exactly 4/5 * the maximum
supported link rate reported in DPCD capabilities. This skews the pixel
clock calculation by exactly that amount, most notably in the case of a
60Hz display modes which incorrectly run at 48Hz.

The following changes have been made to fix this:
- The effective link rate was originally assigned in DpDisplay::Query().
This method now leaves the link rate uninitialized (with a value of 0).
- Introduced a new method called DisplayDevice::InitWithDpllState which
assigns the link rate based on the existing DPLL state. i915::Controller
now calls this on a display before attaching a pipe to it if it
determines that the corresponding DPLL is already powered.
- The maximum link rate selection logic has now been moved to Init(),
which now selects the link rate prior to link training.

Bug: 78470
Test: 1. intel-i915-test
      2. `fx iquery show-file /dev/diagnostics/class/display-controller-impl/000.inspect`
         shows the bootloader-configured link rate for the built-in eDP
         display on atlas 4K and 1080p and the maximum available link
         rate for an external monitor with a DisplayPort connection.
      3. Use `fx shell session_control add http://www.vsynctester.com`
         to verify that atlas now runs at 60Hz on both 1080p and 4K. Do
         this test on both using the internal eDP panel and an external
         DisplayPort monitor using
         `fx shell stash_ctl set i_can_haz_display_id int #ID`
Change-Id: I091498a2df825ce4e1765711e935a2a2bcc05e8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589005
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Arman Uguray <armansito@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Reviewed-by: Craig Stout <cstout@google.com>