[debian_guest] Only authenticate with cipd if not logged in
The `mkcipd.sh` script currently requires the user to set up a password
each time the script is run, preventing automation of the script.

Update the script to only prompt for authentication if the user is not
logged in.

Change-Id: I9b257289641f9df71bc52fcc35f4ef7e1a844df5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592423
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>