[kernel][vm] Check for overflow in eviction target
Check for overflows when adding a new min pages target to the existing
one. In practice this should never happen as it implies we've been asked
to evict an unreasonably large amount of pages. This guards against
garbage values and bugs in target calculation arithmetic.

Change-Id: I611c9b1ee9dc59015dac1550813afe00dbea6a61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592022
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>