[kernel][vm] Unify Imminent-OOM logic with other levels
Imminent-OOM is a diagnostic level, so we did not trigger eviction at
it. The expectation was that eviction would already have been triggered
before we reached Imminent-OOM, and we synchronously evict anyway at
OOM. However if the pressure level is oscillating between Normal and
Imminent-OOM without hitting the intermediate levels, we will only be
relying on synchronous eviction at OOM, leaving the device at the
Imminent-OOM level for longer.

There can be cases where memory spikes suddenly from Normal to OOM,
leaving us in this degenerate case. Use the Imminent-OOM level as a
trigger for eviction too to catch this case. This will let asynchronous
eviction clear up some memory sooner in extreme scenarios where the
system remains at the Imminent-OOM level.

With this change the Imminent-OOM level is no longer a diagnostic level
for the kernel; it is still a diagnostic level for userspace as
memorypressure clients are not notified.

Change-Id: Ice5ed9dde928b4343b91a79ec9382b206e077336
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592024
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>