[kernel][vm] Fix race in ProcessLruQueues
If there are concurrent calls to ProcessLruQueues, then the lru
generation that is read outside of the lock could become stale.
This is a problem since that stale value of the lru would be
re-written to lru_gen_, potentially causing lru_gen_ to move backwards,
which would be an error.

Change-Id: Ie95bce45cd7654c2f8e84991e7fafb9c9a9efa17
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592429
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>