Reland "[a11y] Migrate web semantics tests to RealmBuilder."
This is a reland of bf6ab7ba63e0d8d04fdf7bc784bbbf66bc06defe

The test seems to be flaky on NUCs, possible because of a missing
iwlwifi driver (though unclear if that's the root cause). Consistent
with other integration tests that rely on netstack (e.g.
touch-input-test), we will restrict the semantics integration tests to
AEMU.

TEST: fx test fuchsia-pkg://fuchsia.com/semantics-integration-tests#meta/web_semantics_tests.cm
Bug: 84831
Change-Id: If39c2527bdd70005b488d129aa93091924ac55af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591630
Commit-Queue: Alexander Brusher <abrusher@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>