[wlan][policy] Log MLME event name in disconnect event
This is to prepare for deprecating wlanstack metrics. In wlanstack, we
include the MLME event name in disconnect events in Inspect, but in
Policy, we currently do not. With this change, the MLME event name is
now logged in Policy as well.

Bug: 82654
Fixed: 84892
Test: Refactor + logging change only
Change-Id: I7e3805ea7d153ef5bbc80b69e3727220d1d09f75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585574
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>