[botanist] Move timeout context to beginning of execution.
Previously the timeout wrapped the subcommand run by botanist (usually
testrunner). However, before it runs the subcommand, it does some other
steps like booting up the target. Thus, the -timeout flag did not tell
the real timeout of the botanist execution, but rather the timeout of
the subcommand it ran. This would cause a problem when setting the task
timeout because we'd have to account for the time taken by the other
steps in botanist which may not always be enough.

Bug: 67855
Change-Id: Ib935cf211f7a11ae9df77fea699ea87b08cc2d0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592708
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>