[zircon] Fixed stream read races
There was a race when truncating a VMO whilst reading from a stream
which would result in ZX_ERR_OUT_OF_RANGE errors that shouldn't happen.

This CL addresses this by adding a lock so that they are correctly
sequenced.  This might not be the most performant way to solve this, but
it's simple enough and we probably don't need this case to be
performant.

Fixed: 83803
Change-Id: I1b9f33c0f44a58bbf7be8b49ed6860586820abeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589962
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>