[a11y] Re-enable flaky WebSemanticsTest.HitTest.
This test case had a couple of issues:

(1) Chrome initially represented the "Test 1 2 3..."
static text as a parent node with label "Test 1 2 3..."
and a set of children each representing a single word.
But later, chrome consolidated the children into a single
child with "Test 1 2 3..." as its label. This changed
seemed to cause the parent's bounds to change. I believe
the test case computed the hit test target based on the
original bounds, which resulted in a miss. I modified the
HTML to use single-word static text to help stabilize the
tree across this consolidation step.

(2) The test case previously did not wait for chrome to
receive its device scale factor before proceeding with the
hit test, resulting in a race condition where the test fixture
computed the hit test target coordinates without accounting for
the scale factor, and then chrome performed the hit test using
the scale factor. I addressed this issue by waiting for the device
scale factor to be reflected in the root node's transform.

TEST: fx test fuchsia-pkg://fuchsia.com/semantics-integration-tests#meta/web_semantics_tests.cm
Bug: 60002
Change-Id: I575792c282f337605b9248cac80b56d288f76a17
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590913
Commit-Queue: Alexander Brusher <abrusher@google.com>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>