[ffx] Make self-test cleanup more resiliant
I thought waiting 80ms would be enough to let the daemon exit, this
looks like it is not enough. This instead does a few checks to see if a
daemon can exit on its own, waiting in between checks 50ms. In the end
it then runs pkill to make sure nothing is left running. This *should*
imply that the final kill is only done if something is really getting
leaked.

Bug: 86030
Change-Id: I1f2724b5e88857193e9719bff138efbc33019ae7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590564
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>