[ffx] Add machine stable output for trace plugin
The only subcommand under trace that produces output is list-providers.
This writes the vector that is returned from the GetProviders call more
or less directly as the machine output. I need an intermediate type
because the ProviderInfo type does not implement Serialize or
Deserialize for some reason.

Fixed: 85919
Change-Id: I58c08f835960bffe51f38ebaa9b871ea27dcca48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591081
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>