[a11y] Make navigation listener/controller member variables.
This change makes the navigation listener and controller member
variables of the WebViewProxy class in the web semantics integration
test fixture. It's possible that these objects can cause flakes when
they go out of scope, so hopefully tying their lifespans to that of the
test fixture should resolve the flakes.

TEST: fx test fuchsia-pkg://fuchsia.com/semantics-integration-tests#meta/web_semantics_tests.cm
Bug: 86243
Change-Id: I9431ed93b7115c6db7f51a09bdccab12ca0d5815
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593245
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>